+++ CHANGELOG.md
@@ -11,6 +11,11 @@
   tooling have all been moved to a single `ethereum/` crate.
 * **BACKWARD INCOMPATIBLE:** Dummy node now requires `--storage-backend BACKEND` option.
   Use backend `dummy` for original behavior.
+* **BACKWARD INCOMPATIBLE:** Bump `protobuf` dependency to `2.0`.
+* **BACKWARD INCOMPATIBLE:** Remove unused key manager support from compute node.
+* **BACKWARD INCOMPATIBLE:** Use TLS for compute node gRPC channels.
+* **BACKWARD INCOMPATIBLE:** Separate node/entity key pair arguments and change serialized
+  node key pair format in a backward-incompatible way.
 * Add discrepancy resolution by using backup workers majority vote.
 * Add passing extra arguments to Docker in shell (`--docker-extra-args`).
 * Add `common::futures::retry` which implements retrying futures on failure.
@@ -25,6 +30,7 @@
   using the environment's default executor.
 * Make clients aware of compute committee changes and manage connections.
 * Added AWS DynamoDB-backed storage backend.
+* Backport `SelectAll` stream combinator from futures 0.3.
 
 # 0.1.0